refactor(phase3): decompose setup.ts and improve consolidate/distill god-functions
- setup: split setup.ts into setup/steps/*, legacy-config, providers, semantic-assets, prompt; move deepMergeConfig/isPlainObject to core
- improve/consolidate: split akmConsolidateInner into pipeline stages over a shared context; four op-branches become dispatched handler functions
- improve/distill: split akmDistill into stages; extract the memory->knowledge promotion branch into its own function

22 files changed, 4295 insertions(+), 3443 deletions(-) (net +852, file-boundary tax + mandated unit tests)
